boefjes/kat_nmap/requirements.txt
python-libnmap==0.7.2
